chore: Move class to separate file
It will likely be reused across many writers.
